+++ frontend/src/components/pages/TeamsSelect.tsx
@@ -1,20 +1,15 @@
-import { ChangeEvent, FC, useContext, useEffect, useState } from "react";
-import { Link, useNavigate } from "react-router-dom";
+import { ChangeEvent, FC, useEffect, useState } from "react";
+import { Link } from "react-router-dom";
 import { Button, MenuItem, TextField, Typography } from "@mui/material";
 import Grid2 from "@mui/material/Unstable_Grid2/Grid2";
 import { AxiosRequestConfig, AxiosResponse } from "axios";
 import { axiosInstance } from "../../utils/axios";
-import { Team, TeamsResponse, TeamsSelectResponse } from "../../types";
-import { SnackbarContext } from "../../providers/SnackbarProvider";
+import { Team, TeamsSelectResponse } from "../../types";
 
 const TeamsSelect: FC = () => {
-  const { handleSetSnackbar } = useContext(SnackbarContext);
-  const navigate = useNavigate();
-
   const [teams, setTeams] = useState<Team[]>([]);
   const [teamId, setTeamId] = useState<string>("");
   const [teamName, setTeamName] = useState<string>("");
-  const [newTeamName, setNewTeamName] = useState<string>("");
 
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
     const { value } = event.target;
@@ -29,43 +24,6 @@ const TeamsSelect: FC = () => {
     if (selectTeam) {
       setTeamName(selectTeam.name);
     }
-  };
-
-  const handleTeamsCreate = () => {
-    const teamsCreateOptions: AxiosRequestConfig = {
-      url: "/teams",
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      data: { name: newTeamName },
-    };
-
-    axiosInstance(teamsCreateOptions)
-      .then((res: AxiosResponse<TeamsResponse>) => {
-        console.log(res);
-        handleSetSnackbar({
-          open: true,
-          type: "success",
-          message: "チームを作成しました",
-        });
-        navigate("/sign_up", {
-          state: {
-            teamId: res.data.team.id,
-            teamName: res.data.team.name,
-            isAdmin: true,
-          },
-        });
-      })
-      .catch((err) => {
-        console.log(err);
-        handleSetSnackbar({
-          open: true,
-          type: "error",
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
-          message: `${err.response.data.messages.join("。")}`,
-        });
-      });
   };
 
   const options: AxiosRequestConfig = {
@@ -90,7 +48,7 @@ const TeamsSelect: FC = () => {
         <Typography variant="h4" component="div" gutterBottom>
           所属チームの選択
         </Typography>
-        <Typography variant="subtitle1" component="div">
+        <Typography variant="body1" component="div">
           選択したチームでユーザー登録します。
         </Typography>
       </Grid2>
@@ -114,46 +72,29 @@ const TeamsSelect: FC = () => {
         </TextField>
       </Grid2>
       <Grid2 xs={12}>
-        <Button variant="contained" type="button">
-          <Link
-            style={{ textDecoration: "none", color: "black" }}
-            to="/sign_up"
-            key={teamId}
-            state={{ teamId, teamName, isAdmin: false }}
-          >
-            次へ
-          </Link>
+        <Button
+          variant="contained"
+          component={Link}
+          to="/sign_up"
+          key={teamId}
+          state={{ teamId, teamName, isAdmin: false }}
+        >
+          次へ
         </Button>
       </Grid2>
       <Grid2 xs={12}>
-        <Typography variant="subtitle1" component="div" sx={{ my: 2, ml: 13 }}>
+        <Typography variant="body1" component="div">
           または
         </Typography>
       </Grid2>
       <Grid2 xs={12}>
-        <Typography variant="h4" component="div" gutterBottom>
-          新規チーム作成
-        </Typography>
-        <Typography variant="subtitle1" component="div">
-          新しいチームの管理者としてユーザー登録します。
-        </Typography>
-      </Grid2>
-      <Grid2 xs={12}>
-        <TextField
-          required
-          type="text"
-          inputProps={{ maxLength: 20 }}
-          label="新規チーム名"
+        <Button
+          variant="contained"
           color="secondary"
-          helperText="20文字以内"
-          value={newTeamName}
-          onChange={(event) => setNewTeamName(event.target.value)}
-          sx={{ width: "30ch" }}
-        />
-      </Grid2>
-      <Grid2 xs={12}>
-        <Button type="submit" variant="contained" onClick={handleTeamsCreate}>
-          作成
+          component={Link}
+          to="/sign_up/teams/new"
+        >
+          新規チーム作成
         </Button>
       </Grid2>
     </Grid2>
